Move to/Use OccurrenceCollection instead, filter for published events
