+++ src/onegov/event/collections/events.py
@@ -6,7 +6,6 @@ from datetime import date, timezone
 from datetime import datetime
 from datetime import timedelta
 from icalendar import Calendar as vCalendar
-from lxml import objectify, etree
 
 from onegov.core.collection import Pagination
 from onegov.core.utils import increment_name
@@ -407,73 +406,3 @@ class EventCollection(Pagination):
 
         return self.from_import(items, publish_immediately=True,
                                 future_events_only=future_events_only)
-
-    def as_xml(self):
-        """
-        Returns the events of the given occurrences as xml.
-
-        Format:
-        <events>
-            <event>
-                <id></id>
-                <title></title>
-                <tags></tags>
-                    <tag></tag>
-                <description></description>
-                <start></start>
-                <end></end>
-                <location></location>
-                <price></price>
-                ..
-            </event>
-            <event>
-                ..
-            </event>
-            ..
-        </events>
-
-        :return: str
-        """
-        xml = '<events></events>'
-        root = objectify.fromstring(xml)
-
-        # for ev in EventCollection(self.session).query().all():
-        # for e in self.session.query(Event).all():
-        for e in self.query().all():
-            event = objectify.Element('event')
-            event.id = e.id
-            event.title = e.title
-            txs = tags(e.tags)
-            event.append(txs)
-            event.description = e.description
-            event.start = e.start
-            event.end = e.end
-            event.location = e.location
-            event.price = e.price
-            event.organizer = e.organizer
-            event.event_url = e.external_event_url
-            event.organizer_email = e.organizer_email
-            event.modified = e.last_change
-            root.append(event)
-
-        # remove lxml annotations
-        objectify.deannotate(root, pytype=True, xsi=True, xsi_nil=True)
-        etree.cleanup_namespaces(root)
-
-        return etree.tostring(root, encoding='utf-8', xml_declaration=True,
-                              pretty_print=True)
-
-
-class tags(etree.ElementBase):
-    """
-    Custom class as 'tag' is a member of class Element and cannot be
-    used as tag name.
-    """
-    def __init__(self, tags=()):
-        super().__init__()
-        self.tag = 'tags'
-
-        for t in tags:
-            tag = etree.Element('tag')
-            tag.text = t
-            self.append(tag)
